*: Add a new command emeritus-stats
emeritus-stats gives how many emeritus_reviewers or
emeritus_approvers were added and removed in a given
time frame for a particular repository.

This commit adds the relevant helper functions to calculate
these numbers are print them. This commit also introduces
new types to facilitate this calculation as well as edits
existing types if needed.

The high level algorithm followed is:

- Checkout the repo at `path` at date `from`
- Collect numbers for `emeritus_reviewers`
  and `emeritus_approvers` for `from`
- Checkout the repo at `path` at date `to`
- Collect similar number for `to`
- Calculate the difference between these two

Signed-off-by: Madhav Jivrajani <madhav.jiv@gmail.com>